feature: add MXNet 1.4.1 Dockerfiles (#102)
starting with MXNet 1.4.1, the Python 3 images will be compatible with
a variety of platforms: SageMaker, EKS, EC2. The Python 2 ones, however,
will not, given that Python 2's deprecation nears. That is why
the Dockerfiles in this PR diverge so much between Python 2 and Python 3.
For now, I've left it in the release buildspec to keep
building and releasing Python 3 images as a sanity check,
but in the future we may elect to remove that.